Добавлен контроллер задачи
